+++ ad-tag/source/ts/ads/adUnitPath.test.ts
@@ -1,0 +1,30 @@
+import { expect, use } from 'chai';
+import sinonChai from 'sinon-chai';
+
+import { removeChildId } from './adUnitPath';
+
+// setup sinon-chai
+use(sinonChai);
+
+describe('ad unit path', () => {
+  describe('remove child network id', () => {
+    it('should not change ad unit paths without child network ids', () => {
+      ['/1234567/Travel', '/1234567/Travel/Europe', '/1234567/Travel/Europe,Berlin'].forEach(
+        adUnitPath => {
+          expect(adUnitPath).to.be.equals(removeChildId(adUnitPath));
+        }
+      );
+    });
+
+    it('should remove child network ids', () => {
+      [
+        ['/1234567,1234/Travel', '/1234567/Travel'],
+        ['/1234567,1234/Travel/Berlin', '/1234567/Travel/Berlin'],
+        ['/1234567,1234/Travel/Europe,Berlin', '/1234567/Travel/Europe,Berlin'],
+        ['/1234567,1234,5678/Travel/Europe,Berlin', '/1234567/Travel/Europe,Berlin']
+      ].forEach(([input, expected]) => {
+        expect(removeChildId(input)).to.be.equals(expected);
+      });
+    });
+  });
+});