+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 3.35.0
+
 [GD-3313](https://jira.gutefrage.net/browse/GD-3313) Sanitize `adUnitPath` for `slotName`. Child network id is removed
 [GD-3319](https://jira.gutefrage.net/browse/GD-3319) Add `vlyby` prebid adapter types